qcacld-3.0: Cleanup snoc interrupt controll
remove interupt enable/disable at group level since
in register set 81, these registers are not provided.

don't enable/disable the ce irqs at the qgic
since the interrupts are edge driven, enabling and disabling the interrupt
in the qgic should not be necessary.  Also remove an unneeded read
of a status register.

Change-Id: I2c20ea26c23dc672697d53cd5433d855d9332b13
CRs-Fixed: 997176