[qcacmn] Fix key installation issue for BE offload radios
For big endian based offload radios, the key installed using
WMI_VDEV_INSTALL_KEY_CMDID requires to be code swapped as the copy engine
for big endian platform does byte swap operation on the received key before
sending it down to FW.

It is observed that the big endian conversion is carried out in ol_if_vap.c
for big endian platform specific radios in the driver, that takes care of
converting the key data to the byte swapped form. However, it is observed
that an additional copy is carried out in the WMI layer that is overwriting
the converted key with old key, in the non-swapped form.

This change has the converged WMI driver changes to add the byte swapping
operation for big endian platforms. It will depend on an equivalent change
on the host specific driver component.

IRs Fixed
IR-172411
IR-172932

Change-Id: I3a35f8f22582aa611441567e40c0805d4bc30670