[qcacmn] Fix ATF peer stats info memory corruption during extraction
In the atf peer stats event extraction, the peer stats info - token info
is extracted for each of the peer connected to the AP. The param sent to the
WMI layer is an instance of the structure supporting only token info for a single
peer. However, in the extraction API, we copy the memory buffer for token info
obtained from FW to the buffer instance created in host.

This will work for a single peer but will result in memory corruption if there
are more than a single peer. In order to fix this, we keep the instance as it is
and change the extraction method to enable indexed extraction of the token info
from the event buffer obtained from FW. We re-use the same single instance defined
in the param structure by calling the extraction routine for token info multiple times
based on the number of ATF peers.

IRs Fixed: 172919,173437
CRs Fixed: 1068268
Change-Id: Ib9b1a7a629fa5b242bc3aec65c12fb0f98cc46d8