qcacmn: pdev param to set firmware to use lower chains for tx
New pdev param with which firmware can be set to use
a lower number of chains without doing a reset. In response
to this firmware's ratecontrol will select only rates
which can be sent using a lower number of chains, and set
the h/w tx descpriptor's chain select to whatever value
has been configured using this command.

Change-Id: Ie30d5e6b61a54e73f0553c6b2e9fa0a9bc8af42e
Signed-off-by: Anirban Sirkhell <anirban@codeaurora.org>