qcacmn: Fix KW issue in extraction API for peer and vdev extd stats
wmi_stats_event structure is defined in FW header wmi_unified.h with
variable length data as A_UINT8 data[0]. Klocwork tool perceives this
as a single uint8 and throws zero sized array when we try to access buffer
following data.

A uint8 pointer is assigned to data for this purpose in order to fake
the KW tool and resolve the issue.

Change-Id: I8dfa347726a975a70b5e23778694e3a1764a547b