qcacld-3.0: Fix memory leak and timer leak during driver unload in FTM mode
When the driver is unloaded in the FTM mode there is memory leak,
as part of start_adapter for the FTM the driver registers for
for action frames and it is not freed, this frame registration
is not required for the FTM mode.

Also, there is invalid check of not destroying the memdump timer
for the FTM mode. Memdump intialization is done as part of driver load
irrespective of con_mode, hence destroy the memdump during unload.

CRs-Fixed: 2031300
Change-Id: I87f8a6d4813926228232aa81e4141355632c9739