qcacld-3.0: Move temp scan list to main list once threshold is reached
In case scan is not in progress, then temp scan list is not moved to main
scan list as no scan complete is received. So in case of SAP, this
list can become very big and cause MC thread loop for long time while
finding the duplicate BSS.

To avoid this if scan is not in progress and temp scan list reach a
threshold, move it to main scan cache list.

Change-Id: Ibf7b69d480cb82eb7df8fb14acf152441bf2bb60
CRs-Fixed: 2047199