qcacld-3.0: Add STA-only IPA offload support
IPA offload is not supported for STA-only mode. In AP-STA config, IPA
offload is enabled only when first client is connected to AP. This is
OK for MSM platforms. But for MDM platforms, this is not the case.
There's STA tethering config, where traffic is routed between STA and
USB tethering devices. IPA offload needs to be enabled for STA to
promote performance.

Fix is to introduce a macro that is MDM platform specific and STA-only
IPA support is decided upon driver build time.

Change-Id: If2bc016aa941f2c0651a2669f1169e631208326b
CRs-Fixed: 2115653