qcacld-3.0: Cleanup of hdd_wlan_startup
Cleanup the hdd_wlan_startup, presently lot of configuration
is written to firmware from the startup function which is not
correct with different driver models supported. Move the feature
initialization to hdd_features_init. Also in the probe failure
error path netdev notifier is not getting unregistered.
Leaving stale pointer with the kernel which can lead to a crash.

Unregister the netdevice during the probe failure and cleanup features.

Change-Id: If7fe7c4c0f97e4b4829551840643c4e525696c76
CRs-Fixed: 2090570