qcacld-3.0: Reset the mScanPending flag after cleaning scan queue
As part of hdd_cleanup_scan_queue, all the pending scan are removed
from the queue but mScanPending flag remain true for the adapter.

And thus in stop adapter, driver assume that scan is pending and
driver waits for 5 sec waiting for scan abort.

To fix this set mScanPending to false while removing all scan req
from the hdd scan queue

Change-Id: Ifa72f3840f1dfbbc1a0cc56d237a69fdf960c4c2
CRs-Fixed: 2119520