qcacld-3.0: Abort SAP restart when unloading
As part of driver remove, the station-softAP concurrency work is
flushed, and then all adapters are stopped. This leads to a race window
where new work can be scheduled after the flush, but before the adapters
are stopped. Reject requests to schedule this work during driver unload
to prevent the race condition.

Change-Id: Icae93c4e6dca364e930568fac2964c376dd4fbc6
CRs-Fixed: 2169387