qcacld-3.0: Use request manager for OCB
We are transitioning to the new request manager framework. Change all
of the OCB interfaces which currently use the hdd_ocb_ctxt to this
framework. Note that this framework provides the infrastructure to
pass data from the response thread to the request thread and hence
eliminates the need to maintain transient OCB data in the HDD adapter
struct.

Change-Id: I4151221b81cfcaec25bad79da4aa25164c0f83ad
CRs-Fixed: 2005320