qcacmn: Account for temporary peers created in object manager
The wlan_max_peer count in the object manager currently
includes the count of temporary peers and actual peers
created on the target, and this could be used to limit the
actual numbers of peers created.

The result is that, for a given number of
peers configured the object manager limits
counting temporary peer also and not actual peers alone.

Indicate the count of actual peers and temporary peers
separately, to appropriately limit actual peers created.

Change-Id: Ia1593d702949853ad26756f167988ec70279eb87
CRs-Fixed: