qcacld-3.0: Avoid VDEV start for new interface when roaming in progress
Currently driver allows start_bss on SAP interface when roaming is in
progress on STA interface. This leads to two simultaneous vdev starts
in FW which causes the FW to assert.
Also after connect or start_bss is queued to the WMA to send VDEV start,
a roaming triggered from FW could lead to the same issue.

Add changes to reject the start_bss request for SAP if roaming is in
progress on any STA interface.
Also, when a connect for STA or start_bss for SAP is received and
roaming is not in progress for any STA interface, then roaming is
stopped on all STA interfaces by sending WMI_ROAM_SCAN_MODE_NONE
to FW. Then after the association is completed for STA or start_bss
is completed for SAP, roaming is enabled again on the STA interface.

Change-Id: I3baaffeef3b350e6527660cbac4b79fa4d9f83f0
CRs-Fixed: 2221337