qcacmn: Do not process bangradar command in NON-DFS channel
FW, on receiving the bangradar command from host, suspends the
scheduler/tx-traffic and resumes only after the channel is switched.

For a bangradar in non-DFS channel, host does not send the channel
switch command. Hence, the tx_traffic are not resumed.

When the user issues a bangradar, forward it to FW only if the current
channel is a DFS channel.

Change-Id: Id1f969dd7654e540a1c637632122cf2c6b543123
CRs-Fixed: 2225953