qcacmn: Re-create self AST entry for a peer when it is re-used
If a peer already exists, when peer_create is issued to DP, DP
just reuses the existing peer entry and clears the appropriate 
state including AST entries tied to the peer, which includes 
peer's self AST entry.

Fix this issue by re-creating the AST entry and adding as self 
AST for the peer.

Change-Id: I3a535dc62f420a875230744285155980b854861d