qcacmn: Fix a double free issue for tx descriptor
Tx descriptors that belong to a particular vdev are released in
vdev detach path. Since DP soc is not detached yet, interrupts
are not disabled, so it is possible that host gets completions
for same tx descriptor and it tries to process it again.

Add a NULL check for vdev in tx completion path to avoid
duplicate processing of tx descriptors

Change-Id: I5a62ef4d981dbfd0a5ca7483acf4270145d016be