qcacld-3.0: Set max bss score to scan entry with BSSID hint
BSS score for scan entry with BSSID same as BSSID hint is
not set to the maximum if it is the first scan entry being
added to the candidate AP list used for connection. This
results in connection not getting initiated to particular
BSSID.

Fix is to set bss score to the max value for scan entry
matching the BSSID hint.

Change-Id: Ic9583ff09c049aeb2d014143c7a2fba2d0c339ef
CRs-Fixed: 2276830