qcacmn: Change bss_peer set logic for sta mode
In case of station mode, self peer will have the mac address of the vdev
and bss_peer will have mac address of AP.

During peer create in case of station mode, if peer mac address equals
that of vdev mac address, then the peer is a self peer and bss_peer flag
should not be set for this peer.

When peer map event is received in station mode, we can set bss_peer
flag as 1 without doing any mac address comparison.

Change-Id: Id27477a500c2b653605e6874de68c7ae4ebed4f3
CRs-fixed: 2287393