qcacld-3.0: Check for target ready before sending wmi command
SAP stop is in progress when FW down event is received
from platform driver. Vdev_stop command is not sent to
firmware due to target ready check but same check is not
present for self peer_delete command. This results in
assert in firmware since vdev_stop command is not received
before self peer_delete command.

Fix is to set wmi_stopinprogress as part of wma shutdown
notifier callback.

Change-Id: I59a20a2a8903370943de00a3731f3e44c736e68f
CRs-Fixed: 2302678