qca-wifi: Fix dfs null pointer dereference issue
When a DFS pulse is being processed, a channel change change can happen.
And the channel change initializes the radar filters which means
it frees and allocates the DFS filtering data structures.
This may result in memory violation (NULL Pointer dereference).

To avoid the memory violation lock the filter initialization and
DFS pulse processing.

Change-Id: Ibb90d067273b24f165205610d76709c6cbc36487
CRs-Fixed: 2317455
Signed-off-by: Vignesh U <uvignesh@codeaurora.org>