qcacld-3.0: Reset recovery in progress flag after probe success
Currently driver recovery in progress flag is set when SSR is
triggered. When the SSR is triggered at load time even before
FW ready comes this flag is getting set. This flag is getting
reset in re-init path, as the driver is loading re-init path
will not be triggered and this flag will never be reset.

To address this issue, reset the flag once the probe is
success in both the reinit and load cases.

Change-Id: I209dbc204fdf031850b6480001cef931ef3c6bc2
CRs-Fixed: 2365220