qcacld-3.0: Fix OOB in wlan_hdd_netif_queue_control
Put spinlock when adapter->history_index is getting incremented.
There is possibility of race condition in the increment of
adapter->history_index as there are multiple call paths that
invoke wlan_hdd_netif_queue_control() that are not all within
the same thread.

Since the check is to see if it equals the max value, we can
exceed the bounds of the array on the next invocation of this
function, resulting in a OOB.

Change-Id: I0d127b2e47bee32cd0b8a0281c6213895f3d4960
CRs-Fixed: 2370801