qcacld-3.0: Add PLD ops to notify thermal level change
As a part of a requirement, the platform driver registers with thermal
subsystem as a cooling device. After this registration, the thermal
subsystem notifies the platform driver when there is a change in the
temperature level according to predefined thresholds in the dtsi.
This notification needs to be forwared to WLAN host, on which
appropriate mitigation action can take place.

Add a PLD ops - set_curr_therm_state() using which the platform driver
sends thermal state change notification to WLAN host. The callback
contains state variable to store the current thermal state.

Change-Id: Idac72a8ac8b1d6f2a5a7ed585a8d9e0470bb99a5
CRs-Fixed: 2468796