qcawifi: Remove unnecessary print from rate stats module
When rate tlvs are nor parsed or completion status is not okay,
still we indicate WDI_EVENT_TX_PPDU_DESC for other features
and it results into feeding invalid rate index stats to rate stats
module, which we can not process, which is a valid scenario.
so just remove err print to avoid flooding cosole.

CRs-Fixed: 2451257
Change-Id: I451a68af18e92727802f78e23d94ab6ea31dd5bb