qca-wifi: Unmark PreCAC NOL based on 160MHz PreCAC tree
PreCAC NOL is unmarked based on the 80MHz center frequency. This is
wrong as the PreCAC tree, with the avilablity of full 160MHz Agile engine,
can have a root as high as 160MHz. This will lead to channel not picked
for PreCAC after NOL timeout.

PreCAC NOL should be unmarked based on the center frequency of
the root node's bandwidth.

CRs-Fixed: 2654993
Change-Id: Ie054d725c1b7a2fed47c9f294f3d71c4dd9e8ecb