qcacmn: Use preamble to get ht/vht/he flags
Use preamble type instead of mode specific flags to fill up the radiotap
header. With current implementation, ppdu_info is not memset to zero every
time. Hence, using the ht/vht/he flags to fill up the radiotap header is
not a correct approach, as the radiotap information for the next PPDU might
get filled based on the previous PPDU information.

Also make rs_fcs_err as 0, in case mpdu stitch fails so that it will not use
the previous rs_fcs_err status.

Change-Id: Ic3299469c5b806170dbd8716b7274975e66ca8be