qcacld-3.0: Remove set roaming in progress in wlan_hdd_reassoc_bssid_hint()
roaming_in_progress flag is for LFR3.0 roaming, which will be set when
roam sync indication reported from target,
sme_neighbor_middle_of_roaming() will check the roaming state for
LFR2.0 roaming. So both case don't have to set this flag when call
wlan_hdd_reassoc_bssid_hint().

If we set roaming in progress when BTM happen with LFR2.0 roaming,
we don't clear it currently. So remove redundant set roaming in
progress for this case. It is part of change:
Change-Id: Ide20a30d9d8c20a96d8bba70367b51b8643cae09

Change-Id: I311ae7422d1dbd181e817c0292e2266c4cd5b005
CRs-Fixed: 2760378