qca-wifi: Disable radar detection after radar found on precac chan
When radar is detected on a zero-wait-precac channel, the precac channel
changes but the phyerror buffers for the previous precac channel keep
coming, affecting the newly configured precac channel and creating false
detects.

Consider, a Cascade AP operating in home channel 36VHT80 and the precac
channel is 52VHT80. When the radar is injected on channel 52VHT80, the
precac channel changes to the next precac channel 100VHT80 but still
continues to receive the queued phyerror buffers, causing radar detects
on channel 100VHT80(false detects with respect to channel 100VHT80).

Disable radar detection after radar got detected on the precac channel.

CRs-Fixed: 2812335
Change-Id: I1e06c53f00e596d9142f3aeb25a00b9da088538b