qcacld-3.0: Disabling the channels for 80MHz in regdb has no effect
Though the regulatory is offloaded to Firmware, the driver takes
the channel bandwidth and centre frequency from beacon. This
results in connecting STA to a channel in 80MHz even though this
channel is restricted for 80MHz in regdb.

This change considers the reg rules sent by FW and filters the
proper channel bandwidth and centre frequency, restricting STA
connection as per the provided reg rules.

Change-Id: I8474621a02b3da6926156304b231c2985547de7b
CRs-Fixed: 2873913