qca-wifi: Addressing NULL ptr deref in interCAC
Issue:
Bring up AP on channel 100 HT80 mode with interCAC and RCAC
features enabled. Configure interCAC channel (dfs_precac_inter_chan_freq)
as 36. AP comes up on channel 36 and rCAC runs on channel 100
(dfs_autoswitch_chan). Once rCAC is completed on chan 100,
AP switches from channel 36 to 100. Once this happens,
"dfs_autoswitch_chan" becomes NULL and "dfs_precac_inter_chan_freq" is
reassigned as channel 100. This is to ensure that next time when we
come up on channel 100, we can skip CAC (preCAC completed chan).

Once on channel 100, inject radar. On NOL expiry of channel 100, there is a
panic due to NULL pointer dereference. This is because, on NOL expiry, we
try to switch to switch to interCAC channel if it exists.
"dfs_precac_inter_chan_freq" holds a value but "dfs_autoswitch_chan" is
NULL as the initial phase of interCAC is done.
As "dfs_autoswitch_chan" checked for NULL along with
"dfs_precac_inter_chan_freq" (which is a valid value), the conditional
statement was incorrect and it lead to a NULL pointer
dereference.

Solution:
Modified the conditional statement to check only for dfs_autoswitch_chan
pointer to address the kernel null pointer dereference.
dfs_autoswitch_chan is set to NULL once we come up on that
channel. However, "dfs_precac_inter_chan_freq" is not set to 0
once interCAC channel change is done. The setting of this channel
is retained so that, next time the same channel can be used to come
up without CAC. Hence validating if dfs_precac_inter_chan_freq is 0
is of no purpose.

Moreover, "dfs_precac_inter_chan_freq" is not required to be
validated, as the selection of rCAC frequency does not depend
on inter CAC channel (a.k.a. "dfs_precac_inter_chan_freq").

CRs-Fixed: 2910367
Change-Id: Ide61ed1025ba6e0219489ada1457368b3b7939a8