qcacld-3.0: Add support for disable chan list command on country change
Currently restore channels does not work on country change which
is leading to below issue:
For a country code as per regulatory if any channels are disabled
then during disable chan list the states of these channels is
cached. Now if the country code is changed and those respective
channels are not disabled as per this new CC but the states of
the channels is still cached which was as per previous country.
Now if restore channels is done then those channels are getting
restored according to cached states irrespective of the current
country. This may lead to an issue where non-supported channels
for a specific country can get enabled.
For example if initially country is IT where channels 12 and 13
are enabled and driver receives command to disable these channels
and after that country is changed to US where channel 12 and 13
are not supported now if driver receives command to restore
channel list then previously disabled channels i.e. channels 12
and 13 are also enabled which is not correct.

To address above issue add a change to update the cached channel
states after every country change according to the country
regulatory information.

Change-Id: I709a8b68a113a7fe60c98367857a33205ac3739b
CRs-Fixed: 2935041