qca-wifi: Check current DFS domain to find if ADFS is enabled
The Agile precac enabled check does not consider the DFS domain check.

Consider a case, where the ADFS was enabled for an ETSI country and the
ADFS picks a channel and runs precac. When a country change command is
issued by the user say "US" country, the ADFS is restarted despite the
new country being a non-ETSI country. This is because the precac
enabled check does not consider the DFS domain check.

Add DFS domain check to precac enabled check to prevent such scenarios.

Change-Id: I3c07489c9283d7542e6c6c86bb6d1adc64e77df1
CRs-Fixed: 2994630