qca-wifi: Do not reset agile dfs index
In case of a DBS-SBS HK board, the ADFS should run as follows:
1) Attach the Agile detector the 5G high radio - wifi 0.
2) pick a channel in wifi 0 and start agile cac.
3) once agile cac is completed, attach the agile detector to the
next radio which is 5G low radio - wifi 1 in our case.
4) pick a channel in wifi 1 and start agile cac.
5) Again after agile cac is done move back to wifi 0.

This flow is broken in ADFS SM because, the current agile dfs index is
cleaned up as part of precac clean up that happens for every time
an agile cac is completed. Since current agile dfs index is cleaned up,
every time the agile detector is attached to the first radio that has a
dfs channel to perform adfs. That is,
1) the agile detector gets attached to the 5G high radio - wifi0 and
performs agile cac on all the channels of wifi0 until all the channels
are marked precac done.
2) It will not resume precac on the 5G low radio. A manual vap down
up on the 5G low radio is required to attach the agile detector to wifi0
and start the ADFS SM.

In order to prevent such scenario, the current agile dfs index should not
be cleaned as part of precac clean up for every precac completion.

But not cleaning up will prevent the ADFS from starting because, when
in the init event handler there is a check to see if the current agile
dfs index has been reset or not. The purpose of the check is to ensure
that the Agile DFS is not running on pdev. This purpose is already
solved by the state variable "precac_state_started" and is checked inside
INIT state START event handler. So, the check that checks whether
current agile dfs index has been reset or not is redundant and can be
removed.

These changes will make sure to alternate the agile detector across all the
radios in a Round-Robin manner but the problem that a manual vap down
up is required on the 5G low radio is requred is not solved because the
"agile_precac_active" is set only after receiving the start event
in the init state. "agile_precac_active" whether adfs is supported in
the radio or not. This has to be set based on the agile precac enabled
check when the vap is brought up. This change is being done is converged
code.

Change-Id: I0ea227e6da5e271a8d9cb48b3ae3c4a7d52b0404
CRs-Fixed: 3009739